fix(ui): align phrase-lock check with render guard, unblock doctor lint
confirm.tsx compared phrase with `!= null` for locking the confirm
button but `phrase ?` for rendering the input, so an empty phrase left
the button enabled with no field to type into. Split dialog-manager's
non-component registry into dialog-store.ts so it stops tripping
react-doctor's only-export-components rule.
